test: extract validation tests
