+++ packages/radio-group/test/radio-group.test.js
@@ -1,5 +1,5 @@
 import { expect } from '@esm-bundle/chai';
-import { fixtureSync, nextFrame, nextRender } from '@vaadin/testing-helpers';
+import { fixtureSync, nextFrame } from '@vaadin/testing-helpers';
 import { sendKeys } from '@web/test-runner-commands';
 import sinon from 'sinon';
 import '../vaadin-radio-group.js';
@@ -429,128 +429,6 @@ describe('radio-group', () => {
       const event = changeSpy.firstCall.args[0];
       expect(event.target).to.equal(buttons[0].inputElement);
       expect(event.target.value).to.equal(buttons[0].value);
-    });
-  });
-
-  describe('initial validation', () => {
-    let validateSpy;
-
-    beforeEach(() => {
-      group = document.createElement('vaadin-radio-group');
-
-      const radio = document.createElement('vaadin-radio-button');
-      radio.value = '1';
-      group.appendChild(radio);
-
-      validateSpy = sinon.spy(group, 'validate');
-    });
-
-    afterEach(() => {
-      group.remove();
-    });
-
-    it('should not validate by default', async () => {
-      document.body.appendChild(group);
-      await nextRender();
-      expect(validateSpy.called).to.be.false;
-    });
-
-    it('should not validate when the field has an initial value', async () => {
-      group.value = '1';
-      document.body.appendChild(group);
-      await nextRender();
-      expect(validateSpy.called).to.be.false;
-    });
-
-    it('should not validate when the field has an initial value and invalid', async () => {
-      group.value = '1';
-      group.invalid = true;
-      document.body.appendChild(group);
-      await nextRender();
-      expect(validateSpy.called).to.be.false;
-    });
-  });
-
-  describe('validation', () => {
-    beforeEach(async () => {
-      group = fixtureSync(`
-        <vaadin-radio-group>
-          <vaadin-radio-button label="Button 1" value="1"></vaadin-radio-button>
-          <vaadin-radio-button label="Button 2" value="2"></vaadin-radio-button>
-        </vaadin-radio-group>
-      `);
-      await nextFrame();
-      buttons = [...group.querySelectorAll('vaadin-radio-button')];
-    });
-
-    it('should not set required property by default', () => {
-      expect(group.required).to.be.undefined;
-    });
-
-    it('should pass validation by default when group is not required', () => {
-      expect(group.checkValidity()).to.be.true;
-      expect(group.invalid).to.be.false;
-    });
-
-    it('should not set invalid when group is required and has not lost focus yet', () => {
-      group.required = true;
-      expect(group.checkValidity()).to.be.false;
-      expect(group.invalid).to.be.false;
-    });
-
-    it('should set invalid when radio group is required and validate is called', () => {
-      group.required = true;
-      group.validate();
-      expect(group.invalid).to.be.true;
-    });
-
-    it('should validate on radio button click', () => {
-      group.required = true;
-      group.validate();
-      buttons[1].click();
-      expect(group.invalid).to.be.false;
-    });
-
-    it('should validate on value change', () => {
-      group.required = true;
-      group.validate();
-      group.value = '1';
-      expect(group.invalid).to.be.false;
-    });
-
-    it('should validate and set invalid on focusout when group is required', async () => {
-      group.required = true;
-
-      // Focus on the first radio button.
-      await sendKeys({ press: 'Tab' });
-      // Move focus out of the group.
-      await sendKeys({ down: 'Shift' });
-      await sendKeys({ press: 'Tab' });
-      await sendKeys({ up: 'Shift' });
-
-      expect(group.checkValidity()).to.be.false;
-      expect(group.invalid).to.be.true;
-    });
-
-    it('should fire a validated event on validation success', () => {
-      const validatedSpy = sinon.spy();
-      group.addEventListener('validated', validatedSpy);
-      group.validate();
-
-      expect(validatedSpy.calledOnce).to.be.true;
-      const event = validatedSpy.firstCall.args[0];
-      expect(event.detail.valid).to.be.true;
-    });
-
-    it('should fire a validated event on validation failure', () => {
-      const validatedSpy = sinon.spy();
-      group.addEventListener('validated', validatedSpy);
-      group.required = true;
-      group.validate();
-
-      expect(validatedSpy.calledOnce).to.be.true;
-      const event = validatedSpy.firstCall.args[0];
-      expect(event.detail.valid).to.be.false;
     });
   });
